Dodanie modułu scrollTop i poprawienie nazw plików/funkcji
