+++ src/js/scrollTop.js
@@ -1,0 +1,8 @@
+const scrollTop = () => {
+  window.scrollTo({
+    top: 0,
+    behavior: 'instant',
+  });
+};
+
+export default scrollTop;
